statistical done

+++ src/main/java/org/example/quan_ao_f4k/dto/response/product/ProductDetailDTO.java
@@ -9,5 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ProductDetailDTO {
 	private String productName;
-	private Integer quantity;
+	private Long quantity;
+
 }
